Fix #1:  Replace MathML with TeX
All of the MathML code is replaced by equivalent TeX formulas that are
much easier to read and modify.  Some of the MathML formulas were
incorrect so we looked at the original text doc for the correct
formula and updated them accordingly.

Thus, this fixes #2 as well.

We also updated to MathJax 3 since it supports normal TeX equation
numbering that is used here.