Add \times between fractions
When two fractions are multiplied together, it's hard to see that
there's a gap, so it looks like just one fraction smashed together in
a funny way.  Increasing the spacing helps a little, but let's make
it more obvious by using \times to indicate that the fractions are
multiplied together.