Add equation number to the coefficient formulas
For each filter type, there is a group of 6 equations giving the
coefficients of the filter.  Add an equation number of the group.
Didn't want to add one for each equation because that would be too
many.  And MathJax doesn't support the subequations environment so we
can't easily add a subequation (1a,1b, ect).